requirements.txt
alembic==1.4.0
astroid==2.3.3
Babel==2.8.0
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
Flask==1.1.1
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
mccabe==0.6.1
oauthlib==3.1.0
passlib==1.7.2
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.4.4
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.13
tweepy==3.8.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
